+++ KintaiTeam09/src/main/java/com/example/demo/controller/ExportController.java
@@ -1,21 +1,19 @@
 package com.example.demo.controller;
 
-import org.springframework.web.bind.annotation.GetMapping;
-
 public class ExportController {
 
 	
-	@GetMapping("/topmenu")
-	public String topmenu() {
-		return "topmenu"; // メニュー画面に遷移
-	}
-
-	// 仮実装(ページ遷移のみ)
-	@GetMapping("/top")
-	public String top() {
-	    return "top"; //トップ画面に遷移
-
-	}
+//	@GetMapping("/topmenu")
+//	public String topmenu() {
+//		return "topmenu"; // メニュー画面に遷移
+//	}
+//
+//	// 仮実装(ページ遷移のみ)
+//	@GetMapping("/top")
+//	public String top() {
+//	    return "top"; //トップ画面に遷移
+//
+//	}
 }
