+++ KintaiTeam09/src/main/java/com/example/demo/controller/RegistController.java
@@ -1,7 +1,6 @@
 package com.example.demo.controller;
 
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
 public class RegistController {
@@ -11,17 +10,18 @@ public class RegistController {
 //    }
 	
 	
-	@GetMapping("/topmenu")
-	public String topmenu() {
-		return "topmenu"; // メニュー画面に遷移
-	}
-
-	// 仮実装(ページ遷移のみ)
-	@GetMapping("/top")
-	public String top() {
-	    return "top"; //トップ画面に遷移
-
-	}
+//	@GetMapping("/topmenu")
+//	public String topmenu() {
+//		return "topmenu"; // メニュー画面に遷移
+//	}
+//
+//	// 仮実装(ページ遷移のみ)
+//	@GetMapping("/top")
+//	public String top() {
+//	    return "top"; //トップ画面に遷移
+//
+//	}
+	
 }
